Extracted the TigerPython define libraries code to a helper method.
